combine all map functions and consts to folder
